app/requirements.txt

# Application
Flask==2.3.3
gunicorn==21.2.0

# Base de données
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# Monitoring
prometheus-flask-exporter==0.22.4
prometheus-client==0.19.0

# Développement (optionnel pour tests)
pytest==7.4.3
pytest-flask==1.2.0
python-dotenv==1.0.0
